refact: rename files and mod for more clarity
