+++ benches/receiver.rs
@@ -7,9 +7,9 @@ use vsmtp::{
     config::server_config::ServerConfig,
     mime::mail::BodyType,
     model::mail::{Body, MailContext},
+    receiver::test_helpers::test_receiver,
     resolver::Resolver,
     rules::address::Address,
-    test_helpers::test_receiver,
 };
 
 #[derive(Clone)]
