+++ src/receiver/tests/mail/ko.txt
