+++ src/resolver/mbox_resolver.rs
@@ -16,7 +16,7 @@
 **/
 use crate::{
     config::{log_channel::RESOLVER, server_config::ServerConfig},
-    model::mail::MailContext,
+    smtp::mail::{Body, MailContext},
 };
 
 use super::Resolver;
@@ -42,10 +42,10 @@ impl Resolver for MBoxResolver {
                     let timestamp = timestamp.format("%c");
 
                     let content = match &ctx.body {
-                        crate::model::mail::Body::Raw(raw) => {
+                        Body::Raw(raw) => {
                             format!("From {} {timestamp}\n{raw}\n", ctx.envelop.mail_from)
                         }
-                        crate::model::mail::Body::Parsed(parsed) => {
+                        Body::Parsed(parsed) => {
                             let (headers, body) = parsed.to_raw();
                             format!(
                                 "From {} {timestamp}\n{headers}\n\n{body}\n",
