--- tests/mod.rs
@@ -1,1 +1,0 @@
-mod integration;
